requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

streamlit

langchain_groq

# RAG System Dependencies
# PDF Processing
PyPDF2

# Vector Stores
faiss-cpu
chromadb

# LangChain Community (Vector Store Integrations)
langchain-community

# Document Processing
langchain-text-splitters